Move componentProps to component idx files
